:refactor: migration of Organization to `src/shared`
