+++ api/tests/school/integration/scripts/create-school-learners_test.js
@@ -1,10 +1,10 @@
-import { Organization } from '../../../../lib/domain/models/Organization.js';
 import {
   buildLearners,
   buildSchoolOrganization,
   showSchools,
 } from '../../../../src/school/scripts/create-school-learners.js';
 import { ORGANIZATION_FEATURE } from '../../../../src/shared/domain/constants.js';
+import { Organization } from '../../../../src/shared/domain/models/Organization.js';
 import { logger } from '../../../../src/shared/infrastructure/utils/logger.js';
 import { databaseBuilder, expect, knex, sinon } from '../../../test-helper.js';
