+++ api/tests/tooling/domain-builder/factory/build-organization.js
@@ -1,4 +1,4 @@
-import { Organization } from '../../../../lib/domain/models/Organization.js';
+import { Organization } from '../../../../src/shared/domain/models/Organization.js';
 
 function buildOrganization({
   id = 123,
