+++ api/tests/tooling/domain-builder/factory/build-user-orga-settings.js
@@ -1,5 +1,5 @@
-import { Organization } from '../../../../lib/domain/models/Organization.js';
 import { UserOrgaSettings } from '../../../../lib/domain/models/UserOrgaSettings.js';
+import { Organization } from '../../../../src/shared/domain/models/Organization.js';
 import { User } from '../../../../src/shared/domain/models/User.js';
 
 /*
